# test-cases/pet-care-store/pet-fe/components/portal/resources/pet-store.swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Cellery Pet Store Sample Controller Service",
    "version": "0.1.0",
    "title": "Cellery Pet Store Sample Controller",
    "contact": {
      "email": "dev@wso2.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
  "tags": [
    {
      "name": "accessories",
      "description": "Accessories for your Pets"
    },
    {
      "name": "pets",
      "description": "Everything to do with Pets"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/catalog": {
      "get": {
        "tags": [
          "accessory",
          "list"
        ],
        "summary": "Get the pet accessories in the Catalog",
        "operationId": "getCatalog",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched Catalog",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          },
          "500": {
            "description": "Server Error Occurred"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "order",
          "list"
        ],
        "summary": "Get the orders placed",
        "operationId": "getOrders",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched Orders list",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "500": {
            "description": "Server Error Occurred"
          }
        }
      },
      "post": {
        "tags": [
          "order",
          "create"
        ],
        "summary": "Place a new order",
        "operationId": "createOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully placed the order",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error Occurred"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": [
          "profile",
          "user"
        ],
        "summary": "Get the profile of the currently logged in user",
        "operationId": "getProfile",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched Profile",
            "schema": {
              "type": "#/definitions/Customer"
            }
          },
          "500": {
            "description": "Server Error Occurred"
          }
        }
      },
      "post": {
        "tags": [
          "profile",
          "user"
        ],
        "summary": "Create a user profile for the user",
        "operationId": "createProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successfully created Profile"
          },
          "500": {
            "description": "Server Error Occurred"
          }
        }
      }
    }
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Id of the item"
        },
        "name": {
          "type": "string",
          "description": "Name of the Item"
        },
        "description": {
          "type": "string",
          "description": "Description about the Item"
        },
        "unitPrice": {
          "type": "number",
          "format": "float",
          "description": "Price of a single unit"
        },
        "inStock": {
          "type": "integer",
          "format": "int64",
          "description": "The number of items in Stock"
        }
      }
    },
    "Customer": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "description": "Username of the customer as stored by the system"
        },
        "firstName": {
          "type": "string",
          "description": "First name of the customer"
        },
        "lastName": {
          "type": "string",
          "description": "Last name of the customer"
        },
        "pets": {
          "type": "array",
          "description": "The pets owned by the customer",
          "enum": [
            "Dog",
            "Cat",
            "Hamster"
          ]
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Username of the customer as stored by the system"
        },
        "customer": {
          "type": "string",
          "description": "First name of the customer"
        },
        "items": {
          "type": "array",
          "description": "Items ordered",
          "items": {
            "type": "#/definitions/Item"
          }
        },
        "orderDate": {
          "type": "string",
          "description": "The date on which the items were ordered"
        },
        "deliveryDate": {
          "type": "string",
          "description": "The date on which the items are estimated to be delivered at"
        },
        "deliveryAddress": {
          "type": "string",
          "description": "Address to deliver the items to"
        }
      }
    }
  }
}